+++ service/go/resources/fonts/license.txt
