+++ service/go/service/click_word_captcha_service.go
@@ -4,9 +4,9 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"golang/model/vo"
-	"golang/util"
-	img "golang/util/image"
+	"github.com/TestsLing/aj-captcha-go/model/vo"
+	"github.com/TestsLing/aj-captcha-go/util"
+	img "github.com/TestsLing/aj-captcha-go/util/image"
 	"log"
 )
