+++ service/go/test/ClickWordCaptchaService_test.go
