+++ service/go/test/cache_util_test.go
@@ -1,7 +1,7 @@
 package test
 
 import (
-	"golang/util"
+	"github.com/TestsLing/aj-captcha-go/util"
 	"testing"
 	"time"
 )
