--- service/golang/go.mod
@@ -1,8 +1,0 @@
-module golang
-
-go 1.18
-
-require (
-	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
-	golang.org/x/image v0.0.0-20220617043117-41969df76e82 // indirect
-)
